move user related components into library
