--- apps/web-main/src/app/context/config.ts
@@ -1,8 +1,0 @@
-const config = {
-  activeMode: 'light', // This can be light or dark
-  isSidebarHover: false,
-  isCollapse: 'mini-sidebar',
-  isLanguage: 'en',
-};
-
-export default config;
