+++ apps/web-main/src/components/project/ProjectContributors.tsx
@@ -15,9 +15,9 @@ import {
 } from '@design-system';
 import { ArrowRightIcon } from 'lucide-react';
 import { Placeholder } from '../ui/Placeholder';
+import { useSession } from '@lib-user';
 import { camelCaseToHuman } from '@lib-utils';
 import { useEffect, useState } from 'react';
-import { useSession } from '../../app/context/SessionContext';
 
 export const ProjectContributors = ({
   project,
