+++ libs/lib-user/README.md
@@ -1,0 +1,7 @@
+# lib-user
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test lib-user` to execute the unit tests via [Jest](https://jestjs.io).
