+++ libs/lib-user/project.json
@@ -1,0 +1,9 @@
+{
+  "name": "lib-user",
+  "$schema": "../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "libs/lib-user/src",
+  "projectType": "library",
+  "tags": [],
+  "// targets": "to see all targets run: nx show project lib-user --web",
+  "targets": {}
+}
