+++ libs/lib-user/src/lib/Login.tsx
@@ -1,16 +1,10 @@
 'use client';
 
-import React, { useCallback, useEffect, useState } from 'react';
+import { useCallback, useEffect, useState } from 'react';
 import { SocialLogin } from './SocialLogin';
-import type { Metadata } from 'next';
 import { Button, H4, MainLogo, Separator } from '@design-system';
 import { EmailLogin } from './EmailLogin';
-export const metadata: Metadata = {
-  title: 'Side Login Authentication',
-  description: 'Generated by create next app',
-};
-
-export type LoginVariation = 'create' | 'login';
+import { LoginVariation } from './types';
 
 const EMAIL_AUTH_ENABLED = false;
 
@@ -106,5 +100,3 @@ export const Login = () => {
     </div>
   );
 };
-
-export default Login;
